Wisp: initial import
This is an interpreter for a Lisp-style language that runs as a
WebAssembly module but can also compile to native code.

It has a basic browser-based REPL written in TypeScript with React as
well as a basic command-line REPL.

This initial import has a working interpreter and garbage collector,
but the language is very incomplete.

The interpreter is a small-step evaluator in the "CEK machine" style,
which means it has an explicit representation of the evaluation
context, also known as the continuation.  Continuations are
heap-allocated Lisp objects that can be inspected and serialized.

The heap can be saved as an image file and loaded again.  We currently
don't save and restore the evaluator's state, but this can be fixed
easily; then we could save and restore even in the middle of
executing code.

The garbage collector uses stop & copy with two semispaces.  It is not
generational or fancy in any way, but it works.  Collection is
triggered between evaluation steps when heap usage exceeds
a threshold.

The TypeScript code reads the heap directly, so it needs to be kept in
sync with the basic memory layout defined by the C code.

I will now proceed to implement more of the basic language primitives
like `PROGN`, `COND`, etc, to make it possible to write some
simple programs.