Implement vararg lambda lists like (lambda (x . y) y)
This is similar to the Scheme style of varargs, different from the
Common Lisp style that looks like (lambda (x &rest y) y).

Maybe I'll change it later but I like this style, it's
very intuitive and easy to implement.

There's some yucky stuff here because sometimes we receive the
evaluated parameters in backwards order for efficiency, but when using
varargs we need the list in forward order.