Basic working version of AWAIT and FETCH
This only works in the browser for now.  I don't particularly feel
like implementing FETCH for Unix; an easier thing would be to get a
Node-based version working, or one of the other Wasm runtimes.

Actually it makes sense to use Node at this stage because I can reuse
the TypeScript that currently only runs in the browser.  That's a
great idea; I should do it right now, probably.

OK, after thinking about it, I'm interested in using Deno instead
of Node.  It makes sense to use a sandboxing environment, and it
supports TypeScript.  Let's see how that works out.