Zig: build to WebAssembly, rudimentary glue
Yay, now we can read, eval, and print from the browser.

Unlike with Emscripten, I'm writing the glue code myself for stuff
like passing strings back and forth.  It's a bit of work, but it's
quite nice to have control over this, rather than relying on
Emscripten's blobs of weird stuff.

So we don't have any POSIX layer with file systems and stuff, but I
think that's okay.