Implement pretty printer in Zig
The algorithm is designed with garbage collection in mind, and I first
thought of using the Wisp heap for it, but I thought it would be
interesting to try using Zig allocators in a more efficient way.

However, I ended up just using an arena allocator.  Pretty printing a
big expression will allocate a bunch of memory for the different
possibilities that the algorithm tries, and then it all gets freed at
the same time.  I haven't measured the viability of this approach but
it might be fine.

Otherwise I can imagine some reference counting scheme or
something else.

Anyway the REPL now uses the pretty printer for results.  It will be
very useful for stepping and debugging.