Use StackFallbackAllocator for arglist scanning
This was the most obvious bottleneck when profiling a simple
recursive loop.

Actually it would be cleaner to initialize one stack allocator at the
top of the step function and pass it down.

It's a very nice pattern; I should embrace this and maybe other forms
of fast temporary allocators, arena allocators, etc.