Debugger: display outer dynamic scopes
The terminology and display might be confusing here.  Most debuggers
just show scopes and then you can click on other stack frames to see
their scopes.  We could possibly do this by letting you click the
continuations in the form view.  That would be cool actually, and
probably not so difficult?

Anyway it seems useful in our Lisp to be able to see the full
dynamic scope easily... but maybe only for variables that actually
have dynamic bindings.  And we don't have dynamic bindings yet!