Unify all continuation types into ktx structure
Before, a continuation could be either a `ktx`, a `cap`, or a `duo`.

We used `cap` for delimited continuation prompts, `duo` for
macroexpansion, and `ktx` for everything else: LET, PROGN, IF, and
function argument evaluations...

But `ktx` was already a multipurpose thing whose fields mean different
things in different situations.  The `fun` field determines the
continuation type.  It could already be LET, PROGN, or IF.  Now it can
also be PROMPT or EVAL.

This makes it much more uniform to scan through and copy continuation
slices, etc.