Work on town tape saving
Now when the town starts a core, it saves its initial tape to disk.

One thing we'll want to do soon is to have an API endpoint that starts
a core, sends its initial tape, and then keeps streaming the
core's log whenever anything happens.

Oh... I just realized that would be easier to do without involving the
tape format, because we can just use a boot log instead.  Then the
tapes are like compacted logs or snapshots -- basically redundant.