Lisp function for showing continuations
This is the same logic that we have in the web code for turning a
continuation inside out and showing it as an expression with a hole
somewhere in the middle... but in Lisp, returning an S-expression.

So we can use this in the REPL, etc.  But we can also later get rid of
the TypeScript version of this logic and just have this version.