Implement a REPL function in Lisp (plus jets like READ, etc)
So now we have some super basic jets for reading and writing to
standard I/O streams, and we expose READ and EVAL, so we can implement
a REPL in Lisp.  The code is of course much more concise than the
Zig REPL!

Now I want the Lisp REPL function to be able to evaluate expressions
inside an error handling context.  This is currently very primitive in
the Zig REPL.  I want to connect the error system with the
continuation prompt system so that we can actually catch errors, etc.