Implement MACROEXPAND and COMPILE (as Lisp functions)
We call it COMPILE because it makes functions and macros faster -- but
actually it just does recursive macroexpansion.  Still, this is a big
and necessary improvement.

During boot, we now also redefine DEFUN to use COMPILE, and then
compile everything that's already been defined.  This takes a while
and allocates a lot of memory.

We really need to fix the garbage collector.  There's a problem with
how it misses some roots when called automatically.