Simple general JavaScript FFI enough to run window.alert
I introduced a new pointer tag type, ext.

An ext contains an index and a value.  The index is a number given to
us from JavaScript, which points out a JavaScript object.  The value
is just some arbitrary metadata that we can use for various purposes.

When the garbage collector runs, it runs a JavaScript release function
for every stale ext.